Switch animation scale to int
This is only ever equal to 1 or 7, as far as I can see.

+++ src/animation.cpp
@@ -38,7 +38,7 @@ namespace
 class basic_animation
 {
     public:
-        basic_animation( const long scale ) :
+        basic_animation( const int scale ) :
             delay{ 0, get_option<int>( "ANIMATION_DELAY" ) * scale * 1000000l } {
         }
